Fill in the detour implementations for the speed hack example
* Update the Interop QueryPerformanceCounter signature to return the Interop.BOOL type instead of a bool

+++ src/Common/src/Interop/Windows/kernel32/Interop.GetTickCount64.cs
@@ -7,6 +7,6 @@ internal partial class Interop
     internal partial class Kernel32
     {
         [DllImport(Libraries.Kernel32)]
-        internal static extern uint GetTickCount64();
+        internal static extern ulong GetTickCount64();
     }
 }
